neutrino: Enable ALLOC_BUFFERS_IN_4K_CHUNKS
Based on kerneltoast/android_kernel_google_wahoo@e321716

With this option disabled, Ion allocates memory in order-4 sized chunks
(16 pages at a time), which is quite a heavy burden on the page
allocator, and can take a while to allocate when memory is fragmented
after a lot of uptime. To lower the allocation latency in Ion, have it
only allocate in page-sized chunks.

[YaroST12: Add missing Kconfig entry]

Co-authored-by: Yaroslav Furman <yaro330@gmail.com>
Co-authored-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: alk3pInjection <webmaster@raspii.tech>